Migrate all to benchmarks folder and make TE import check work
